Agregendo la docuemntación correspondiente al Hash table
Cambios realizados:
- Se agrego la documentación del Hash Table
- Se comenzo a generar la arquitectura de directorios del proyecto
