+++ p1-odProgram.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include "hash_table.c"
+#include "lib/structures/hash_table.c"
 
 int main(void) 
 {
